Move compatibility helpers any, all, minkey, iff to the core.
