+++ sympy/utilities/__init__.py
@@ -1,15 +1,7 @@
 """Some utilities that may help.
 """
-import sys
-
-from iterables import iff, flatten, subsets, variations, \
+from iterables import iff, flatten, subsets, variations, any, all, \
                       numbered_symbols, capture
-
-if sys.version_info[0] <= 2 and sys.version_info[1] < 5:
-    from iterables import any, all
-else:
-    any = any
-    all = all
 
 from lambdify import lambdify
 from source import source
@@ -17,5 +9,3 @@ from source import source
 from decorator import threaded, deprecated
 
 from cythonutils import cythonized
-
-del sys
